refactor(web): compose Resume workbench panes
